refactoring to fragments from activity, WIP.
