+++ Moodlet/app/src/main/java/com/cmput3owo1/moodlet/fragments/MoodHistoryFragment.java
@@ -1,5 +1,6 @@
 package com.cmput3owo1.moodlet.fragments;
 
+import android.content.Intent;
 import android.os.Bundle;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -11,9 +12,11 @@ import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 
 import com.cmput3owo1.moodlet.R;
+import com.cmput3owo1.moodlet.activities.MoodEditorActivity;
 import com.cmput3owo1.moodlet.adapters.MoodEventAdapter;
 import com.cmput3owo1.moodlet.models.MoodEvent;
 import com.cmput3owo1.moodlet.services.MoodEventService;
+import com.google.android.material.floatingactionbutton.FloatingActionButton;
 
 import java.util.ArrayList;
 
@@ -24,6 +27,7 @@ public class MoodHistoryFragment extends Fragment
     private MoodEventAdapter recyclerAdapter;
     private ArrayList<MoodEvent> moodEventList;
     private MoodEventService moodEventService;
+    private FloatingActionButton addMood;
 
 
     public View onCreateView(@NonNull LayoutInflater inflater,
@@ -42,12 +46,26 @@ public class MoodHistoryFragment extends Fragment
         moodEventService = new MoodEventService();
         moodEventService.getMoodHistoryUpdates(this);
 
+        addMood = view.findViewById(R.id.add_mood_fab);
+        addMood.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+                Intent intent = new Intent(getActivity(), MoodEditorActivity.class);
+                intent.putExtra("add",true);
+                startActivity(intent);
+            }
+        });
+
         return view;
     }
+
 
     @Override
     public void onItemClick(int pos) {
         // Implement for editing a mood event
+        Intent intent = new Intent(getActivity(), MoodEditorActivity.class);
+        intent.putExtra("view",true);
+        startActivity(intent);
     }
 
     @Override
