+++ Moodlet/app/src/main/res/layout/fragment_mood_history.xml
@@ -1,4 +1,3 @@
-  
 <?xml version="1.0" encoding="utf-8"?>
 <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
